Move pair_plus_device() to connection
